pyproject.toml
[tool.poetry]
name = "test_project"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.8.10"
pandas = "1.5.2"
fastapi = "^0.88.0"
uvicorn = {extras = ["standart"], version = "^0.20.0"}
pydantic = "^1.10.2"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
mysql-connector-python = "^8.0.33"
python-multipart = "^0.0.6"
install = "^1.3.5"
psycopg2-binary = "^2.9.6"
pika = "^1.3.2"
tarantool = "^1.0.0"
requests = "^2.31.0"
redis = "^4.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
